Better align initWithValue: and createWithValue: for array input
Init required an exact match for number of values, while create
(implicitly) required that the array have no more elements than
properties and that the primary key (if any) be present. Change init to
match create as it's the backwards-compatible way to make them the same,
and improve the errors thrown by create for the wrong number of values.